requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.16
django-cors-headers==4.5.0
djangorestframework==3.15.2
pyodbc==5.2.0
python-decouple==3.8
pytz==2024.2
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.6.0
python-dotenv==1.0.1